pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "alembic==1.16.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "bcrypt==4.3.0",
    "celery==5.4.0",
    "certifi==2025.1.31",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==44.0.2",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi==0.115.11",
    "fastapi-cli==0.0.7",
    "flower==2.0.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "minio==7.2.15",
    "packaging==25.0",
    "passlib==1.7.4",
    "pluggy==1.5.0",
    "psutil>=7.0.0",
    "psycopg2==2.9.10",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pycryptodome==3.22.0",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings==2.8.0",
    "pygments==2.19.1",
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "redis==5.2.1",
    "rich==13.9.4",
    "rich-toolkit==0.13.2",
    "rsa==4.9",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.39",
    "sqlmodel==0.0.24",
    "starlette==0.46.1",
    "typer==0.15.2",
    "typing-extensions==4.12.2",
    "urllib3==2.4.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchfiles==1.0.4",
    "websockets==15.0.1",
]
